Developer guideline exercise:
There is a Magazine which is publishing every month and it has several subscribers.
This code designed based on Observer Behavioral Pattern. It contains three components, such as,
•	MagazinePublisher class,
•	two Observer classes
•	A protocol named Observeral.
This code able to add and remove observers (subscribers) from the array and call the update method by notifyObserver method to notify each observer that a state (event) is changed. BusinessLogic method is responsible for complete the event and call the notifyObserver method.